requirements.txt
streamlit
requests
newspaper3k
lxml_html_clean
google-generativeai
Pillow
numpy